+++ src/pages/leehyeongjoon/Main/Feed/Feed.scss
@@ -1,70 +1,3 @@
-body {
-  margin: 0;
-}
-
-nav {
-  display: flex;
-  justify-content: center;
-  height: 60px;
-  border: 1px solid rgb(232 232 232);
-  background-color: rgb(255 255 255);
-
-  .nav-wrapper {
-    display: flex;
-    justify-content: space-around;
-    width: 1130px;
-
-    .title-wrapper {
-      width: 190px;
-      display: flex;
-      justify-content: space-evenly;
-
-      .insta-icon {
-        width: 30px;
-        height: 30px;
-        padding-top: 12px;
-      }
-
-      .title-line {
-        width: 2px;
-        height: 40px;
-        margin-top: 10px;
-        background-color: black;
-      }
-
-      .title {
-        font-family: 'Lobster', cursive;
-        font-size: 30px;
-        margin-top: 4px;
-      }
-    }
-
-    .input-wrapper {
-      margin-top: 18px;
-
-      .input-box {
-        width: 175px;
-        height: 20px;
-        border: 1px solid #dbdbdb;
-        background: fafafa;
-        background: #fafafa;
-        border-radius: 3px;
-      }
-    }
-
-    .icon-wrapper {
-      display: flex;
-      margin-top: 18px;
-
-      .nav-icon {
-        width: 30px;
-        height: 30px;
-        padding-left: 10px;
-      }
-    }
-  }
-}
-
 main {
   display: flex;
   justify-content: center;
@@ -173,30 +106,30 @@ main {
       }
     }
 
-    // .main-reply {
-    //   display: flex;
-    //   flex-direction: column;
-    //   margin-top: 10px;
+    .main-reply {
+      display: flex;
+      flex-direction: column;
+      margin-top: 10px;
 
-    //   .reply {
-    //     display: flex;
-    //     padding: 3px 0 0 10px;
+      .reply {
+        display: flex;
+        padding: 3px 0 0 10px;
 
-    //     .reply-id {
-    //       font-weight: bold;
-    //     }
+        .reply-id {
+          font-weight: bold;
+        }
 
-    //     .reply-info {
-    //       font-size: 15px;
-    //       padding: 0 5px 0 5px;
-    //     }
+        .reply-info {
+          font-size: 15px;
+          padding: 0 5px 0 5px;
+        }
 
-    //     .see {
-    //       font-size: 15px;
-    //       color: rgb(153 153 153);
-    //     }
-    //   }
-    // }
+        .see {
+          font-size: 15px;
+          color: rgb(153 153 153);
+        }
+      }
+    }
     .main-reply-box {
       display: flex;
